Remove utf8_codepoint_at
